requirements.txt
# Core for trading bot
python-binance==1.0.19
ccxt==4.2.80

# Data processing
pandas==2.2.2
numpy==1.26.4

# Environment & logging
python-dotenv==1.0.1
schedule==1.2.1
loguru==0.7.2
pyyaml==6.0.2

# AI / OpenAI API
openai==1.42.0
